refactor: consolidate all snapshot concerns into snapshot.js
- snapshot.js: 970 lines — snapshotPreprocess(), addSnapshotPlayer(),
  snapshotPoloidal(), snapshotPoloidalPreview(), snapshotSpectrum(),
  plus all internal WebGL helpers (drawPoloidalDataPlotly,
  drawPoloidalDataWebGL, buildShaderProgram, createColorMap, etc.)
- plot-data-process.js: 1118 → 338 lines — now contains only
  historyMode, trackingPlot, addSimulationRegion, cal_gamma,
  cal_omega_r, cal_spectrum
- index.js: 505 → 430 lines — addSnapshotPlayer now accepts
  openPanel/getDataThenPlot as callbacks instead of reaching into
  module scope; snapshotPoloidalPreview import moved to snapshot.js
- Delete snapshot-preprocess.js (consolidated into snapshot.js)
