+++ CLAUDE.md
@@ -54,17 +54,18 @@ This is an Express.js app that visualizes GTC (Gyrokinetic Toroidal Code) simula
 
 ### Client-side (`public/`)
 
-- **`javascripts/index.js`** — Main module: figure lifecycle (openPanel, getDataThenPlot), bootstrap wiring, tab switching. Imports `state.js`, `util.js`, `api.js`, `units.js`, `figure-range-controls.js`, `status-bar.js`, and `navigation.js`.
+- **`javascripts/index.js`** — Main module: figure lifecycle (openPanel, getDataThenPlot, cleanPlot, cleanPanel), bootstrap wiring, tab switching. Imports `state.js`, `util.js`, `api.js`, `units.js`, `figure-range-controls.js`, `status-bar.js`, `navigation.js`, `download.js`, `history-recal.js`, and `snapshot.js`.
 - **`javascripts/status-bar.js`** — Application shell: `StatusBar` class (info/warn/err display), `getStatusBar()`, `wrap()` (async error handling), and `addLoadingIndicator()` (loading spinner wrapper). Used by virtually every other module; depends only on well-known DOM elements.
 - **`javascripts/navigation.js`** — Breadcrumb bar and copy-path button setup: fetches the server-side file tree on load, constructs dropdown folder lists for each breadcrumb segment, manages expand/collapse and click-outside-to-close behaviour. Called once from the `load` handler.
 - **`javascripts/download.js`** — GTC output file download form: `setupDownloadForm()` wires the expand/collapse file-list button and the "Download all/selected" submit buttons. Delegates the POST request to `api.downloadOutputFiles`.
 - **`javascripts/history-recal.js`** — History-mode recalculate button: `addHistoryRecal(panel)` appends a "Recalculate growth rate and frequency" button that re-fits using the user's zoom range and re-renders the affected figures.
+- **`javascripts/snapshot.js`** — All snapshot concerns: `snapshotPreprocess()` dispatches to spectrum/poloidal pipelines; `addSnapshotPlayer()` drives frame-by-frame animation; `snapshotPoloidal()` / `snapshotPoloidalPreview()` / `snapshotSpectrum()` handle data processing and rendering (Plotly carpets + WebGL preview with FFTW-based spectrum analysis).
 - **`javascripts/state.js`** — Centralized application state object (replaces ad-hoc `window.GTCGlobal` properties). Documented with owning concern per property block (units, figure lifecycle, history mode, parameters, FFT). Also exposed as `window.GTCGlobal` for backward compatibility with modules not yet migrated to ES imports.
 - **`javascripts/util.js`** — Shared pure-utility functions with no app-state or DOM dependencies: `callEventTarget`, `propagateFetchError`, `nodeIs`, `postForm`, `min_max`, `interleave`, `unInterleave`.
 - **`javascripts/api.js`** — Server communication layer: `requestPlotData()` fetches plot data/metadata from the Express backend; `downloadOutputFiles()` POSTs a download request and returns the blob + filename. All fetch URL construction and error propagation lives here.
 - **`javascripts/units.js`** — Time-unit conversion: `getBasicParameters()` fetches/caches simulation params; `refreshTimeUnitFactor()` recomputes conversion factors from base unit R₀/c_s to R₀/v_A, tstep, and μs; `applyTimeUnitToFigures()` rescales History/RadialTime x-axis values. Exports `TIME_UNIT_LABEL` for Plotly axis titles.
 - **`javascripts/figure-range-controls.js`** — Per-panel collapsible "Figure Range" controls: `ensurePlotRangeControls()` creates the container, `renderPlotRangeControls()` builds per-figure X/Y min/max input forms, `refreshPlotRangeControls()` syncs inputs from live Plotly axis ranges. Two-way sync via `plotly_relayout` events; also feeds history-mode zoom ranges into `state.hist_mode_range`.
-- **`javascripts/plot-data-process.js`** — Client-side computation: growth rate fitting for history mode data, FFT-based spectrum analysis for snapshots, poloidal plane preview rendering. Imports array utilities from `util.js`.
+- **`javascripts/plot-data-process.js`** — Client-side computation: growth rate fitting for history mode data (`historyMode`, `cal_gamma`, `cal_omega_r`, `cal_spectrum`), particle tracking (`trackingPlot`), equilibrium simulation region overlay (`addSimulationRegion`). Imports array utilities from `util.js`.
 - **`javascripts/summary-generate.js`** — Renders the summary page from equilibrium data
 - **`javascripts/drop-down.js`** — Folder browser dropdown interaction (loaded by `index.pug`, separate from the plot page)
 - **`javascripts/plotly-custom.min.js`** — Custom Plotly.js bundle (built from `custom-plotly/`)
