+++ public/javascripts/plot-data-process.js
@@ -1,8 +1,11 @@
 'use strict';
 
-import { interleave, unInterleave, min_max } from './util.js';
-import { getStatusBar } from './status-bar.js';
+// Snapshot functions (snapshotSpectrum, snapshotPoloidalPreview,
+// snapshotPoloidal, drawPoloidalDataPlotly, drawPoloidalDataWebGL,
+// buildShaderProgram, createColorMap, packTextureArgs, getTicks,
+// getRationalSurface) have moved to snapshot.js
 
+import { interleave, unInterleave } from './util.js';
 export async function historyMode(
     figures,
     interval1 = null,
@@ -121,709 +124,6 @@ export async function historyMode(
         );
     }
 }
-
-export async function snapshotSpectrum(figures) {
-    const field = figures.pop().extraData;
-    const torNum = field.length;
-    const polNum = field[0].length;
-
-    const mmodes = Math.floor(polNum / 5);
-    const nmodes = Math.floor(torNum / 5);
-
-    const modulo = (re, im) => Math.sqrt(re * re + im * im);
-
-    const poloidalSpectrum = Array(mmodes).fill(0);
-    const planPol = new fftw['r2c']['fft1d'](polNum);
-    for (let section of field) {
-        const powerSpectrum = planPol.forward(section);
-        poloidalSpectrum[0] += powerSpectrum[0];
-        for (let i = 1; i < mmodes; i++) {
-            poloidalSpectrum[i] +=
-                2 * modulo(powerSpectrum[2 * i], powerSpectrum[2 * i + 1]);
-        }
-    }
-    planPol.dispose();
-
-    function transpose(matrix) {
-        let result = new Array(matrix[0].length);
-        for (let i = 0; i < result.length; i++) {
-            result[i] = matrix.map(line => line[i]);
-        }
-        return result;
-    }
-
-    const toroidalSpectrum = Array(nmodes).fill(0);
-    const planTor = new fftw['r2c']['fft1d'](torNum);
-    for (let section of transpose(field)) {
-        const powerSpectrum = planTor.forward(section);
-        toroidalSpectrum[0] += powerSpectrum[0];
-        for (let i = 1; i < nmodes; i++) {
-            toroidalSpectrum[i] +=
-                2 * modulo(powerSpectrum[2 * i], powerSpectrum[2 * i + 1]);
-        }
-    }
-    planTor.dispose();
-
-    figures[0].data[0].x = [...Array(mmodes).keys()];
-    figures[0].data[0].y = poloidalSpectrum.map(
-        v => Math.sqrt(v / torNum) / polNum
-    );
-
-    figures[1].data[0].x = [...Array(nmodes).keys()];
-    figures[1].data[0].y = toroidalSpectrum.map(
-        v => Math.sqrt(v / polNum) / torNum
-    );
-}
-
-export async function snapshotPoloidalPreview(figures) {
-    const { polNum, radNum } = figures.pop();
-
-    await drawPoloidalDataWebGL(document.querySelector('#poloidal-preview'), {
-        radNum,
-        polNum,
-        x: figures[0].data[0].x,
-        y: figures[0].data[0].y,
-        z: figures[0].data[1].z,
-    });
-}
-
-export async function snapshotPoloidal(figures, safetyFactor, quick, playing) {
-    const MIN_PTS = 10;
-    const { polNum, radNum } = figures.pop();
-
-    const flattenedField = figures[0].data[1].z;
-    const diagFluxLineColor = 'rgba(142.846, 176.35, 49.6957, 0.9)';
-    const diagFlux =
-        GTCGlobal.basicParameters.diag_flux ?? GTCGlobal.basicParameters.iflux;
-
-    if (!playing) {
-        drawPoloidalDataPlotly(
-            figures[0],
-            radNum,
-            polNum,
-            diagFlux,
-            diagFluxLineColor
-        );
-    }
-
-    if (quick) {
-        return;
-    }
-
-    // calculate spectrum profile on radial grids
-
-    const selectedPoloidalModeNum = [
-        ...new Set(window.GTCGlobal.basicParameters.mmodes),
-    ];
-    const modeNum = selectedPoloidalModeNum.length;
-    if (Math.floor(polNum / MIN_PTS) < Math.max(...selectedPoloidalModeNum)) {
-        getStatusBar().warn = 'm modes in gtc.in is too high!';
-    }
-
-    const spectrumFigureData = [];
-    for (let i = 0; i < 3 * modeNum; i++) {
-        spectrumFigureData.push({
-            y: [],
-            name: `m = ${selectedPoloidalModeNum[Math.floor(i / 3)]}, ${
-                i % 3 == 0 ? 'real' : i % 3 == 1 ? 'imag' : 'modulus'
-            }`,
-            showlegend: true,
-            hoverinfo: 'none',
-            visible: i % 3 == 2,
-            max_: -Infinity,
-            min_: Infinity,
-        });
-    }
-
-    if (!window.GTCGlobal.fftPlan) {
-        const planConstructor = fftw['r2c']['fft1d'];
-        window.GTCGlobal.fftPlan = new planConstructor(polNum);
-    }
-    const plan = window.GTCGlobal.fftPlan;
-
-    const extra_spectrum_data = Array.from(
-        { length: polNum / MIN_PTS },
-        (_, i) => {
-            return {
-                y: [],
-                name: `m = ${i}`,
-                showlegend: false,
-                hoverinfo: 'name',
-                visible: false,
-                max_: -Infinity,
-            };
-        }
-    );
-    for (let r = 0; r < radNum; r++) {
-        const circle = flattenedField.slice(r * polNum, (r + 1) * polNum);
-        plan.forward(circle).forEach((amp, i) => {
-            const mode_num = Math.floor(i / 2);
-            if (mode_num < extra_spectrum_data.length) {
-                const extra_trace = extra_spectrum_data[mode_num];
-                if (i % 2 == 0) {
-                    extra_trace.y.push(amp);
-                } else {
-                    const mod = Math.sqrt(
-                        Math.pow(amp, 2) + Math.pow(extra_trace.y.pop(), 2)
-                    );
-                    extra_trace.y.push(mod);
-                    extra_trace.max_ = Math.max(extra_trace.max_, mod);
-                }
-            }
-            if (!selectedPoloidalModeNum.includes(mode_num)) {
-                return;
-            }
-            const trace_index =
-                3 * selectedPoloidalModeNum.indexOf(mode_num) + (i % 2);
-            const trace = spectrumFigureData[trace_index];
-            trace.y.push(amp);
-            if (amp > trace.max_) {
-                trace.max_ = amp;
-            }
-            if (amp < trace.min_) {
-                trace.min_ = amp;
-            }
-            if (i % 2 == 1) {
-                const modulus = Math.sqrt(
-                    Math.pow(spectrumFigureData[trace_index - 1].y.at(-1), 2) +
-                        Math.pow(trace.y.at(-1), 2)
-                );
-                const modulus_trace = spectrumFigureData[trace_index + 1];
-                modulus_trace.y.push(modulus);
-                if (modulus > modulus_trace.max_) {
-                    modulus_trace.max_ = modulus;
-                }
-                if (modulus < modulus_trace.min_) {
-                    modulus_trace.min_ = modulus;
-                }
-            }
-        });
-    }
-
-    extra_spectrum_data
-        .sort((a, b) => b.max_ - a.max_)
-        .forEach((fig, ind) => {
-            if (ind < 8) {
-                fig.showlegend = true;
-            }
-        });
-
-    let min_values = [Infinity, Infinity, Infinity];
-    let max_values = [-Infinity, -Infinity, -Infinity];
-    spectrumFigureData.forEach((trace, ind) => {
-        max_values[ind % 3] =
-            trace.max_ > max_values[ind % 3] ? trace.max_ : max_values[ind % 3];
-        min_values[ind % 3] =
-            trace.min_ < min_values[ind % 3] ? trace.min_ : min_values[ind % 3];
-    });
-
-    spectrumFigureData.push(...extra_spectrum_data);
-
-    const extend_range = (a, b) => [1.1 * a - 0.1 * b, -0.1 * a + 1.1 * b];
-    const limits = extend_range(
-        Math.max(...max_values),
-        Math.min(...min_values)
-    );
-
-    const RS_POINT_NUM = 20;
-    if (playing) {
-        for (let i = 0; i < GTCGlobal.rational_surface_count + 1; i++) {
-            const vl = figures[1].data[i];
-            if (i == 0) {
-                vl.y = limits;
-                continue;
-            }
-            vl.y = Array.from({ length: RS_POINT_NUM }).map(
-                (_, i) =>
-                    limits[0] +
-                    ((limits[1] - limits[0]) * i) / (RS_POINT_NUM - 1)
-            );
-        }
-    } else {
-        // add rational surface
-        const rational_surface = safetyFactor
-            ? getRationalSurface(
-                  safetyFactor,
-                  window.GTCGlobal.basicParameters.nmodes,
-                  window.GTCGlobal.basicParameters.mmodes
-              )
-            : [];
-        GTCGlobal.rational_surface_count = rational_surface.length;
-        spectrumFigureData.unshift(
-            ...rational_surface.map(({ n, m, r }) => {
-                const pos = window.GTCGlobal.basicParameters.mpsi * r;
-                return {
-                    name: `${n},${m} surface`,
-                    x: Array(RS_POINT_NUM).fill(pos),
-                    y: Array.from({ length: RS_POINT_NUM }).map(
-                        (_, i) =>
-                            limits[0] +
-                            ((limits[1] - limits[0]) * i) / (RS_POINT_NUM - 1)
-                    ),
-                    mode: 'lines',
-                    showlegend: false,
-                    hoverinfo: 'name',
-                    line: {
-                        color: diagFluxLineColor,
-                        dash: 'dash',
-                        width: 1,
-                    },
-                };
-            })
-        );
-        // add diag flux indicator
-
-        spectrumFigureData.unshift({
-            name: 'Diagnostic Flux',
-            x: [
-                GTCGlobal.basicParameters.diag_flux,
-                GTCGlobal.basicParameters.diag_flux,
-            ],
-            y: limits,
-            mode: 'lines',
-            showlegend: true,
-            hoverinfo: 'none',
-            line: {
-                color: diagFluxLineColor,
-                width: 3,
-            },
-        });
-    }
-    // add control buttons
-    // traces: diag flux | rational surfaces | selected m modes (real, imag, modulus) | some largest m modes
-    const pre_len = 1 + GTCGlobal.rational_surface_count;
-    const step3_pick = i =>
-        Array.from(
-            spectrumFigureData,
-            (_, ind) =>
-                ind < pre_len ||
-                ((ind - pre_len) % 3 == i &&
-                    ind < pre_len + 3 * selectedPoloidalModeNum.length)
-        );
-    if (playing) {
-        figures[1].data.splice(
-            1 + GTCGlobal.rational_surface_count,
-            Infinity,
-            ...spectrumFigureData
-        );
-    } else {
-        figures[1].data = spectrumFigureData;
-    }
-    figures[1].layout.updatemenus = [
-        {
-            x: 0.05,
-            xanchor: 'left',
-            y: 0.9,
-            yanchor: 'top',
-            buttons: [
-                ...[2, 0, 1].map(i => {
-                    return {
-                        method: 'update',
-                        args: [
-                            { visible: step3_pick(i) },
-                            {
-                                'xaxis.range': [
-                                    0,
-                                    window.GTCGlobal.basicParameters.mpsi,
-                                ],
-                                'yaxis.range': extend_range(
-                                    min_values[i],
-                                    max_values[i]
-                                ),
-                            },
-                        ],
-                        label: ['Even Parity', 'Odd Parity', 'Modulus'][i],
-                    };
-                }),
-                ,
-                {
-                    method: 'update',
-                    args: [
-                        {
-                            visible: Array.from(
-                                spectrumFigureData,
-                                (_, ind) =>
-                                    ind <
-                                        1 + GTCGlobal.rational_surface_count ||
-                                    ind >=
-                                        1 +
-                                            GTCGlobal.rational_surface_count +
-                                            3 * selectedPoloidalModeNum.length
-                            ),
-                        },
-                        {
-                            'xaxis.range': [
-                                0,
-                                window.GTCGlobal.basicParameters.mpsi,
-                            ],
-                            'yaxis.range': extend_range(
-                                0,
-                                extra_spectrum_data[0].max_
-                            ),
-                        },
-                    ],
-                    label: 'Full',
-                },
-            ],
-        },
-    ];
-    figures[1].layout.yaxis.range = extend_range(min_values[2], max_values[2]);
-}
-
-function drawPoloidalDataPlotly(
-    figure,
-    rad_num,
-    pol_num,
-    diag_flux,
-    diag_line_color
-) {
-    // add carpet indices
-
-    const theta_mesh = [];
-    const psi_mesh = [];
-    for (let r = 0; r < rad_num; ++r) {
-        for (let p = 0; p <= pol_num; ++p) {
-            theta_mesh.push(p);
-            psi_mesh.push(r);
-        }
-    }
-
-    figure.data[0].a = theta_mesh;
-    figure.data[0].b = psi_mesh;
-    figure.data[1].a = theta_mesh;
-    figure.data[1].b = psi_mesh;
-
-    // draw diagnostic flux indicator
-
-    const diagnostic_flux_line = {
-        name: 'Diagnostic Flux',
-        mode: 'lines',
-        line: {
-            color: diag_line_color,
-            width: 3,
-            shape: 'spline',
-            smoothing: 1,
-        },
-        hoverinfo: 'none',
-        type: 'scatter',
-        showlegend: true,
-        x: figure.data[0].x.slice(
-            diag_flux * (pol_num + 1),
-            (diag_flux + 1) * (pol_num + 1)
-        ),
-        y: figure.data[0].y.slice(
-            diag_flux * (pol_num + 1),
-            (diag_flux + 1) * (pol_num + 1)
-        ),
-    };
-
-    figure.data.push(diagnostic_flux_line);
-}
-
-/**
- * @param {HTMLDivElement} container
- * @param {{radNum: number, polNum:number, x:[number], y:[number], z:[number]}} data
- */
-async function drawPoloidalDataWebGL(container, data) {
-    // create canvas
-
-    const create_canvas = id => {
-        const canvas = document.createElement('canvas');
-        canvas.id = id;
-        canvas.width = canvas.height = 700;
-        return canvas;
-    };
-    const figure_canvas = (id =>
-        document.getElementById(id) ?? create_canvas(id))('pol-canvas');
-    const overlay_canvas = (id =>
-        document.getElementById(id) ?? create_canvas(id))('pol-canvas-overlay');
-    if (container.childElementCount == 0) {
-        container.append(figure_canvas, overlay_canvas);
-    }
-
-    // webgl2 context
-    const gl = figure_canvas.getContext('webgl2', { alpha: true });
-    if (!gl) {
-        throw 'Your browser do not support webgl2!';
-    }
-
-    // 2d context for frame an ticks
-    const ctx = overlay_canvas.getContext('2d');
-    ctx.font = '24px serif';
-
-    const { radNum, polNum, x, y, z } = data;
-    // find bounding box
-    const [x_min, x_max] = min_max(x.slice((radNum - 1) * (polNum + 1)));
-    const [y_min, y_max] = min_max(y.slice((radNum - 1) * (polNum + 1)));
-    const z_range = min_max(z);
-
-    // create color map texture
-
-    // Sampling from ColorData["ThermometerColors"]
-    // const color_map_data = new Uint8Array([
-    //     41, 30, 202, 87, 113, 234, 151, 191, 241, 203, 226, 225, 229, 217, 188,
-    //     224, 168, 137, 191, 91, 86, 136, 21, 42,
-    // ]);
-    // Blend between blue and red (not pure blue and red though, they are actually cf[0] and cf[1], with
-    //  cf = ColorData["TemperatureMap"])
-    const color_map_data = new Uint8Array([
-        45, 77, 238, 97, 122, 242, 150, 166, 246, 202, 210, 250, 255, 255, 255,
-        243, 199, 201, 231, 144, 148, 220, 89, 95, 208, 34, 41,
-    ]);
-    const legend_width = 0.1 * (x_max - x_min);
-    const legend_left_padding = 0.1 * (x_max - x_min);
-
-    // bounding box should be aware of color map
-
-    const x_max_total = x_max + legend_left_padding + legend_width;
-    const bounding_box = {
-        center: [0.5 * (x_min + x_max_total), 0.5 * (y_min + y_max)],
-        // add padding through slightly stretch the bounding box dimension
-        dim: [1.05 * (x_max_total - x_min), 1.05 * (y_max - y_min)],
-        z_range,
-    };
-
-    const color_map = createColorMap(
-        gl,
-        ctx,
-        color_map_data,
-        [x_max + legend_left_padding, y_min + 0.1 * (y_max - y_min)],
-        [legend_width, 0.8 * (y_max - y_min)],
-        bounding_box,
-        z_range
-    );
-
-    // create shader program
-    const shader_program =
-        container.gl_shader ??
-        buildShaderProgram(gl, [
-            {
-                type: gl.VERTEX_SHADER,
-                code: await (await fetch('/shader/pol.vert')).text(),
-            },
-            {
-                type: gl.FRAGMENT_SHADER,
-                code: await (await fetch('/shader/pol.frag')).text(),
-            },
-        ]);
-
-    container.gl_shader = shader_program;
-
-    // set uniforms
-    gl.useProgram(shader_program);
-    for (const [key, val] of Object.entries(bounding_box)) {
-        gl.uniform2f(gl.getUniformLocation(shader_program, key), ...val);
-    }
-    gl.uniform2f(
-        gl.getUniformLocation(shader_program, 'resolution'),
-        figure_canvas.width,
-        figure_canvas.height
-    );
-    // bind texture sampler
-    gl.uniform1i(gl.getUniformLocation(shader_program, 'color_map'), 0);
-
-    const grid_num = radNum * (polNum + 1);
-    const color_map_vertex_num = 2 * (color_map_data.length / 3 - 1);
-    const grid_coords = new Float32Array((grid_num + color_map_vertex_num) * 3);
-    for (let i = 0; i < grid_num; ++i) {
-        grid_coords[i * 3] = x[i];
-        grid_coords[i * 3 + 1] = y[i];
-        grid_coords[i * 3 + 2] = z[i];
-    }
-    const element_num = 2 * (polNum + 1) * (radNum - 1);
-    const triangle_stride_vertex_indices = new Uint32Array(element_num);
-    for (let r = 0; r < radNum - 1; ++r) {
-        for (let p = 0; p <= polNum; ++p) {
-            const idx = r * (polNum + 1) + p;
-            triangle_stride_vertex_indices[2 * idx] = idx;
-            triangle_stride_vertex_indices[2 * idx + 1] = idx + (polNum + 1);
-        }
-    }
-
-    // create vertex array object for contour
-
-    const VAO = gl.createVertexArray();
-    gl.bindVertexArray(VAO);
-
-    // create buffers, initialize their data store
-
-    const VBO = gl.createBuffer();
-    gl.bindBuffer(gl.ARRAY_BUFFER, VBO);
-    gl.bufferData(gl.ARRAY_BUFFER, grid_coords, gl.STREAM_DRAW);
-
-    const EBO = gl.createBuffer();
-    gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, EBO);
-    gl.bufferData(
-        gl.ELEMENT_ARRAY_BUFFER,
-        triangle_stride_vertex_indices,
-        gl.STREAM_DRAW
-    );
-
-    // set vertex attribute
-
-    gl.enableVertexAttribArray(0);
-    gl.vertexAttribPointer(0, 3, gl.FLOAT, 0, 0, 0);
-
-    // specify color map texture
-    color_map.bindToTextureUnit(gl.TEXTURE0);
-
-    // Clear canvas with white background color
-    gl.clearColor(0, 0, 0, 0);
-    gl.clear(gl.COLOR_BUFFER_BIT);
-
-    // draw call
-    gl.bindVertexArray(VAO);
-    gl.drawElements(gl.TRIANGLE_STRIP, element_num, gl.UNSIGNED_INT, 0);
-    color_map.draw();
-}
-
-/**
- *
- * @param {WebGL2RenderingContext} gl
- */
-function buildShaderProgram(gl, shader_info) {
-    const program = gl.createProgram();
-
-    function compileShader(type, code) {
-        const shader = gl.createShader(type);
-
-        gl.shaderSource(shader, code);
-        gl.compileShader(shader);
-
-        if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-            console.log(
-                `Error compiling ${
-                    type === gl.VERTEX_SHADER ? 'vertex' : 'fragment'
-                } shader:`
-            );
-            console.log(gl.getShaderInfoLog(shader));
-        }
-        return shader;
-    }
-
-    shader_info.forEach(({ type, code }) => {
-        const shader = compileShader(type, code);
-
-        if (shader) {
-            gl.attachShader(program, shader);
-        }
-    });
-
-    gl.linkProgram(program);
-
-    if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
-        console.log('Error linking shader program:');
-        console.log(gl.getProgramInfoLog(program));
-    }
-
-    return program;
-}
-
-/**
- *
- * @param {WebGL2RenderingContext} gl
- * @param {CanvasRenderingContext2D} ctx For drawing ticks
- * @param {Uint8Array} data
- * @param {[number, number]} corner lower left corner
- * @param {[number, number]} dim width and height
- * @param {{center:[number, number], dim:[number, number], z_range:[number, number]}} bounding_box
- * @param {[number, number]} z_range
- */
-function createColorMap(gl, ctx, data, corner, dim, bounding_box, z_range) {
-    const color_map_texture = gl.createTexture();
-    gl.bindTexture(gl.TEXTURE_2D, color_map_texture);
-    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.NEAREST);
-    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.NEAREST);
-
-    gl.texImage2D(...packTextureArgs(gl, data));
-
-    // As far as I'm using another draw call for color map, creating separate vbo/ebo
-    //  for it should not be a performance issue.
-
-    const VAO = gl.createVertexArray();
-    gl.bindVertexArray(VAO);
-    gl.bindBuffer(gl.ARRAY_BUFFER, gl.createBuffer());
-    gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, gl.createBuffer());
-
-    const color_num = data.length / 3;
-
-    // geometry thingy
-
-    const [x0, y0] = corner;
-    const x1 = x0 + dim[0];
-    const dy = dim[1] / color_num;
-
-    const [z0, z1] = z_range;
-    const dz = (z1 - z0) / color_num;
-
-    const vertex_data = new Float32Array(3 * 4 * color_num);
-    const element_data = new Int32Array(3 * 2 * color_num);
-    for (let i = 0; i < 2 * color_num; ++i) {
-        vertex_data[3 * (2 * i)] = x0;
-        vertex_data[3 * (2 * i) + 1] = y0 + Math.floor((i + 1) / 2) * dy;
-        vertex_data[3 * (2 * i) + 2] = z0 + (Math.floor(i / 2) + 0.5) * dz;
-
-        vertex_data[3 * (2 * i + 1)] = x1;
-        vertex_data[3 * (2 * i + 1) + 1] = y0 + Math.floor((i + 1) / 2) * dy;
-        vertex_data[3 * (2 * i + 1) + 2] = z0 + (Math.floor(i / 2) + 0.5) * dz;
-
-        const idx = Math.floor((4 * Math.floor((i * 3) / 2)) / 3);
-        element_data[3 * i] = idx;
-        element_data[3 * i + 1] = idx + 1;
-        element_data[3 * i + 2] = idx + 2;
-    }
-
-    // initial data store for buffers
-
-    gl.bufferData(gl.ARRAY_BUFFER, vertex_data, gl.STREAM_DRAW);
-    gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, element_data, gl.STREAM_DRAW);
-
-    // set vertex attribute
-
-    gl.enableVertexAttribArray(0);
-    gl.vertexAttribPointer(0, 3, gl.FLOAT, 0, 0, 0);
-
-    return {
-        /**
-         * @param {number} texture
-         */
-        bindToTextureUnit: function (texture) {
-            gl.activeTexture(texture);
-            gl.bindTexture(gl.TEXTURE_2D, color_map_texture);
-        },
-        draw: function () {
-            this.bindToTextureUnit(gl.TEXTURE0);
-            gl.bindVertexArray(VAO);
-            gl.drawElements(gl.TRIANGLES, color_num * 6, gl.UNSIGNED_INT, 0);
-
-            ctx.textBaseline = 'middle';
-            const ticks = getTicks(z_range, color_num);
-            const canvas_width = ctx.canvas.width;
-            const canvas_height = ctx.canvas.height;
-            ctx.clearRect(0, 0, canvas_width, canvas_height);
-
-            const {
-                center: [cx, cy],
-                dim: [w, h],
-            } = bounding_box;
-            const tick_x =
-                5 +
-                (x1 - cx) * Math.min(canvas_width / w, canvas_height / h) +
-                0.5 * canvas_width;
-
-            ticks.forEach((tick, idx) => {
-                const tick_y =
-                    canvas_height -
-                    ((y0 + (idx + 1) * dy - cy) *
-                        Math.min(canvas_width / w, canvas_height / h) +
-                        0.5 * canvas_height);
-                const label = tick.toExponential(2);
-                ctx.fillText(label, tick_x, tick_y);
-            });
-        },
-    };
-}
-
 export async function trackingPlot(figures) {
     const zeta = figures.pop().extraData;
 
@@ -1036,79 +336,3 @@ export function cal_spectrum(reals, images, timeStep, interval) {
 
 // min_max is now imported from util.js
 
-/**
- * @param {WebGL2RenderingContext} gl
- * @param {Uint8Array} data
- */
-function packTextureArgs(gl, data) {
-    return [
-        gl.TEXTURE_2D,
-        0,
-        gl.RGB,
-        data.length / 3,
-        1,
-        0,
-        gl.RGB,
-        gl.UNSIGNED_BYTE,
-        data,
-    ];
-}
-
-function getTicks([min, max], num) {
-    const get_exp = x => Math.floor(Math.log10(Math.abs(x)));
-    const get_mantissa = (x, n) => x * Math.pow(10, -(n ?? get_exp(x)));
-
-    const tweak_mantissa = (x, f, n) =>
-        f(get_mantissa(x, n)) * Math.pow(10, n ?? get_exp(x));
-
-    const truncated_mid_pt = tweak_mantissa(
-        0.5 * (max + min),
-        Math.floor,
-        Math.max(get_exp(max), get_exp(min))
-    );
-
-    const step = tweak_mantissa(
-        (2 * Math.max(truncated_mid_pt - min, max - truncated_mid_pt)) / num,
-        x => 0.04 * Math.floor(25 * x)
-    );
-
-    return Array.from(
-        { length: num - 1 },
-        (_, idx) => (idx - 0.5 * (num - 2)) * step
-    );
-}
-
-function getRationalSurface(safetyFactor, n_modes, m_modes) {
-    const mode_num = n_modes
-        .map((n, i) => {
-            return { n: n, m: m_modes[i] };
-        })
-        .sort((a, b) => a.m / a.n - b.m / b.n)
-        .filter(function (item, pos, ary) {
-            const last_item = ary[!pos ? 0 : pos - 1];
-            return !pos || item.m * last_item.n != item.n * last_item.m;
-        });
-    const linear_map = (t, x0, x1, y0, y1) =>
-        y0 + ((y1 - y0) * (t - x0)) / (x1 - x0);
-
-    const result = [];
-    const [r0, r1] = window.GTCGlobal.basicParameters.radial_region;
-    for (let i = 0; i < safetyFactor.x.length - 1; ++i) {
-        if (safetyFactor.x[i + 1] < r0 || safetyFactor.x[i] > r1) {
-            continue;
-        }
-        mode_num.forEach(mode => {
-            const pos = linear_map(
-                mode.m / mode.n,
-                safetyFactor.y[i],
-                safetyFactor.y[i + 1],
-                safetyFactor.x[i],
-                safetyFactor.x[i + 1]
-            );
-            if (pos >= safetyFactor.x[i] && pos < safetyFactor.x[i + 1]) {
-                result.push({ r: (pos - r0) / (r1 - r0), ...mode });
-            }
-        });
-    }
-    return result;
-}
